Fix org#2704 about old FD compatibility
Old client (prior to 11.0.0) job messages are silently dropped from
the director logs. The Job status is correct, but the messages are
not displayed in the catalog or the bacula-dir.log.

The workaround is to enable the local log on each impacted FD,
or upgrade them to 13.0.x.

This fix will detect the client version and apply a fix dynamically
when getting a message.